Installed and configured eslint, added support for dotenv, added script to run dev server, general cleanup
