+++ .env.example
@@ -1,0 +1,3 @@
+SERVER_PORT=2500
+SERVER_URL=http://localhost
+MONGO_URL=mongodb://localhost/testdb
