+++ package.json
@@ -1,18 +1,35 @@
 {
-  "name": "server",
+  "name": "judo-in-cloud-be",
   "version": "1.0.0",
   "description": "",
-  "main": "dist/server.js",
+  "main": "dist/main.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "cd typescript; tsc; cd ../dist; node server.js"
+    "start": "ts-node src/main.ts",
+    "start-dev": "tsnd --respawn src/main.ts",
+    "lint": "eslint --fix --ext .ts,.js ."
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@types/express": "^4.17.13",
     "cors": "^2.8.5",
+    "dotenv": "^16.0.1",
     "express": "^4.18.0",
+    "mongoose": "^6.3.3",
+    "ts-node": "^10.7.0",
     "typescript": "^4.6.3"
+  },
+  "devDependencies": {
+    "@types/cors": "^2.8.12",
+    "@types/express": "^4.17.13",
+    "@types/mongoose": "^5.11.97",
+    "@typescript-eslint/eslint-plugin": "^5.23.0",
+    "@typescript-eslint/parser": "^5.23.0",
+    "eslint": "^8.15.0",
+    "eslint-config-standard": "^17.0.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-n": "^15.2.0",
+    "eslint-plugin-promise": "^6.0.0",
+    "ts-node-dev": "^1.1.8"
   }
 }
