+++ src/routers/tournament_router.ts
@@ -1,0 +1,7 @@
+import express = require('express');
+/** api for tournaments */
+export const tournament_router = express.Router();
+
+tournament_router.get('/test', async (req, res) => {
+  res.json({ success: 1 });
+});
